Chat: Correctly update the channel
The chat channel should be updated every time the patch properties change.
